wron dir fix

+++ deployment/deploy.sh
@@ -1,5 +1,7 @@
 #!/bin/env bash
 
-docker rmi $(docker images -f "dangling=true" -q)
+if [[ $(docker images -f "dangling=true" -q) ]]; then
+    docker rmi $(docker images -f "dangling=true" -q)
+fi
 
 docker-compose up --build -d
